Don't try to call s.indexOf unless we know s is a string
We had an issue where we were calling console.error() with an object, and line 26 here was causing node to totally throw up with an uncaught exception error. Calling console.error with an object is perfectly legitimate (https://developer.mozilla.org/en-US/docs/Web/API/Console/error), and oh-ten-bc shouldn't get in the way of that.